+++ src/main/java/enums/ResponseAttribute.java
@@ -1,4 +1,4 @@
-package Enums;
+package enums;
 
 /**
  *
