+++ src/main/java/enums/SessionAttribute.java
@@ -1,4 +1,4 @@
-package Enums;
+package enums;
 
 /**
  *
